Extract TextCol to pricing components
