+++ src/pricing/components/index.js
@@ -1,3 +1,4 @@
 export * from './PricingCol';
 export * from './PricingColChildren';
 export * from './PricingRow';
+export * from './TextCol';
