+++ src/pricing/components/lib/index.js
@@ -1,0 +1,1 @@
+export * from './featurePricing';
